Remove all unnecessary at this point files
