+++ src/Auth/Common/SecurityHeadersAttribute.cs
@@ -1,11 +1,10 @@
 ﻿// Copyright (c) Brock Allen & Dominick Baier. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
 
-
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Filters;
 
-namespace Codidact.Auth.Quickstart.UI
+namespace Codidact.Auth.Common
 {
     public class SecurityHeadersAttribute : ActionFilterAttribute
     {
