+++ src/Auth/Models/ErrorViewModel.cs
@@ -1,10 +1,9 @@
 ﻿// Copyright (c) Brock Allen & Dominick Baier. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
 
-
 using IdentityServer4.Models;
 
-namespace Codidact.Auth.Quickstart.UI
+namespace Codidact.Auth.Models
 {
     public class ErrorViewModel
     {
