+++ src/Auth/Models/LoginViewModel.cs
@@ -1,11 +1,10 @@
 // Copyright (c) Brock Allen & Dominick Baier. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
 
-
-namespace Codidact.Auth.Quickstart.UI.Device
+namespace Codidact.Auth.Models
 {
-    public class DeviceAuthorizationInputModel : ConsentInputModel
+    public class LoginViewModel : LoginInputModel
     {
-        public string UserCode { get; set; }
+        public bool AllowRememberLogin { get; set; } = true;
     }
 }
