+++ src/Auth/Models/RedirectViewModel.cs
@@ -1,11 +1,10 @@
 // Copyright (c) Brock Allen & Dominick Baier. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
 
-
-namespace Codidact.Auth.Quickstart.UI
+namespace Codidact.Auth.Models
 {
-    public class LogoutViewModel : LogoutInputModel
+    public class RedirectViewModel
     {
-        public bool ShowLogoutPrompt { get; set; } = true;
+        public string RedirectUrl { get; set; }
     }
 }
