--- src/Auth/Quickstart/Consent/ConsentOptions.cs
@@ -1,16 +1,0 @@
-﻿// Copyright (c) Brock Allen & Dominick Baier. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
-
-
-namespace Codidact.Auth.Quickstart.UI
-{
-    public class ConsentOptions
-    {
-        public static bool EnableOfflineAccess = true;
-        public static string OfflineAccessDisplayName = "Offline Access";
-        public static string OfflineAccessDescription = "Access to your applications and resources, even when you are offline";
-
-        public static readonly string MustChooseOneErrorMessage = "You must pick at least one permission";
-        public static readonly string InvalidSelectionErrorMessage = "Invalid selection";
-    }
-}
