--- src/Auth/Quickstart/Device/DeviceAuthorizationViewModel.cs
@@ -1,12 +1,0 @@
-// Copyright (c) Brock Allen & Dominick Baier. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
-
-
-namespace Codidact.Auth.Quickstart.UI.Device
-{
-    public class DeviceAuthorizationViewModel : ConsentViewModel
-    {
-        public string UserCode { get; set; }
-        public bool ConfirmUserCode { get; set; }
-    }
-}
