--- src/Auth/Views/Device/UserCodeCapture.cshtml
@@ -1,14 +1,0 @@
-@model string
-
-<div class="page-header">
-    <h1>
-        User Code
-    </h1>
-    <p>
-        Please enter the code displayed on your device
-    </p>
-    <form asp-action="UserCodeCapture" method="post">
-        <input for="userCode" name="userCode" />
-        <button class="btn btn-primary">Submit</button>
-    </form>
-</div>
