--- src/Auth/Views/Diagnostics/Index.cshtml
@@ -1,32 +1,0 @@
-@model DiagnosticsViewModel
-
-<h1>Authentication cookie</h1>
-
-<h3>Claims</h3>
-<dl>
-    @foreach (var claim in Model.AuthenticateResult.Principal.Claims)
-    {
-        <dt>@claim.Type</dt>
-        <dd>@claim.Value</dd>
-    }
-</dl>
-
-<h3>Properties</h3>
-<dl>
-    @foreach (var prop in Model.AuthenticateResult.Properties.Items)
-    {
-        <dt>@prop.Key</dt>
-        <dd>@prop.Value</dd>
-    }
-</dl>
-
-@if (Model.Clients.Any())
-{
-    <h3>Clients</h3>
-    <ul>
-        @foreach (var client in Model.Clients)
-        {
-            <li>@client</li>
-        }
-    </ul>
-}
