+++ src/Auth/Views/Shared/_Layout.cshtml
@@ -31,7 +31,7 @@
                 <a href="~/">
                     <span class="navbar-brand">
                         <img src="~/icon.png" class="icon-banner">
-                        IdentityServer4
+                        Codidact
                     </span>
                 </a>
             </div>
