+++ src/Auth/Views/_ViewImports.cshtml
@@ -1,2 +1,2 @@
-﻿@using Codidact.Auth.Quickstart.UI
+﻿@using Codidact.Auth.Models
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
